+++ webrtc/common_audio/real_fourier.cc
@@ -31,7 +31,7 @@ rtc::scoped_ptr<RealFourier> RealFourier::Create(int fft_order) {
 
 int RealFourier::FftOrder(int length) {
   CHECK_GT(length, 0);
-  return WebRtcSpl_GetSizeInBits(length - 1);
+  return WebRtcSpl_GetSizeInBits(static_cast<uint32_t>(length - 1));
 }
 
 int RealFourier::FftLength(int order) {
