+++ webrtc/modules/audio_coding/codecs/ilbc/xcorr_coef.c
@@ -55,11 +55,11 @@ int WebRtcIlbcfix_XcorrCoef(
 
   /* Find scale value and start position */
   if (step==1) {
-    max=WebRtcSpl_MaxAbsValueW16(regressor, (int16_t)(subl+searchLen-1));
+    max=WebRtcSpl_MaxAbsValueW16(regressor, subl + searchLen - 1);
     rp_beg = regressor;
     rp_end = &regressor[subl];
   } else { /* step==-1 */
-    max=WebRtcSpl_MaxAbsValueW16(&regressor[-searchLen], (int16_t)(subl+searchLen-1));
+    max=WebRtcSpl_MaxAbsValueW16(&regressor[-searchLen], subl + searchLen - 1);
     rp_beg = &regressor[-1];
     rp_end = &regressor[subl-1];
   }
