+++ webrtc/modules/audio_coding/codecs/isac/fix/source/bandwidth_estimator.c
@@ -374,7 +374,7 @@ int32_t WebRtcIsacfix_UpdateUplinkBwImpl(BwEstimatorstr *bweStr,
         /* compute inverse receiving rate for last packet, in Q19 */
         numBytesInv = (uint16_t) WebRtcSpl_DivW32W16(
             524288 + ((pksize + HEADER_SIZE) >> 1),
-            pksize + HEADER_SIZE);
+            (int16_t)(pksize + HEADER_SIZE));
 
         /* 8389 is  ~ 1/128000 in Q30 */
         byteSecondsPerBit = (uint32_t)(arrTimeDiff * 8389);
