+++ webrtc/modules/audio_coding/main/test/initial_delay_unittest.cc
@@ -144,7 +144,7 @@ class InitialPlayoutDelayTest : public ::testing::Test {
     acm_b_->SetInitialPlayoutDelay(initial_delay_ms);
     while (rms < kAmp / 2) {
       in_audio_frame.timestamp_ = timestamp;
-      timestamp += in_audio_frame.samples_per_channel_;
+      timestamp += static_cast<uint32_t>(in_audio_frame.samples_per_channel_);
       ASSERT_GE(acm_a_->Add10MsData(in_audio_frame), 0);
       ASSERT_EQ(0, acm_b_->PlayoutData10Ms(codec.plfreq, &out_audio_frame));
       rms = FrameRms(out_audio_frame);
